Soft delete de produtos e consulta de produtos por categoria

+++ src/main/java/br/com/fiap/lanchonete/application/adapters/controllers/ProdutoController.java
@@ -1,12 +1,20 @@
 package br.com.fiap.lanchonete.application.adapters.controllers;
 
-import br.com.fiap.lanchonete.domain.adapters.services.ProdutoServiceImpl;
-import br.com.fiap.lanchonete.domain.dtos.ProdutoDto;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import br.com.fiap.lanchonete.domain.adapters.services.ProdutoServiceImpl;
+import br.com.fiap.lanchonete.domain.dtos.ProdutoDto;
 
 @RestController
 @RequestMapping("/lanchonete/v1/produtos")
@@ -28,14 +36,22 @@ public class ProdutoController {
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
+    @GetMapping("/categoria/{idCategoria}")
+    public ResponseEntity<List<ProdutoDto>> getProdutosByCategoria(@PathVariable Integer idCategoria) {
+        return ResponseEntity.ok(produtoServiceImpl.findByCategoria(idCategoria));
+    }
+
     @GetMapping
     public ResponseEntity<List<ProdutoDto>> getAllProdutos() {
-        List<ProdutoDto> produtos = produtoServiceImpl.findAll();
+        List<ProdutoDto> produtos = produtoServiceImpl.findAllByAtivoTrue();
         return ResponseEntity.ok(produtos);
     }
 
     @DeleteMapping("/{idProduto}")
     public ResponseEntity<Void> deleteProdutoById(@PathVariable String idProduto) {
+        if (produtoServiceImpl.findByIdProduto(idProduto).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         produtoServiceImpl.deleteByIdProduto(idProduto);
         return ResponseEntity.ok().build();
     }
